feat(backend): Implementation of HTTPS for backend
Utilizing the previously created `server.key` and `server.crt` files
stored in the `certs` directory. All of these are not included in the
git index because they should be created locally and not shared without
encription. The sharing is outside the scope of the repo, and they
should never be published. Even if encrypted, on the repo.