chore(frontend)!: moved crm files to subdirectory
Due to having to use both a CRM and a donations page for the users, the
frontend directory must both include the CRM project, and the donaciones
project. This change only adds the CRM project subdirectory.

BREAKING CHANGE: This is ultimately a total redesign of the file
structure and must be taken seriously.