refactor(backend): Moved JWT handling to module
Now the JWT generation and checking will be handled in the module,
instead of having to be in the `index.js` file. For better code
organization. The function is meant to be called before the backend
grants access to the endpoint. To verify that the request was given
with an appropriate JWT.