refactor(backend): Move resource control methods to separate files
The control methods which interact with the database are separated into
their own files, improving modularity and ensuring that the
implementation details are left separate from the routing functionality.
Hence the routes have only limited access to the database and don't
interact with it directly. Making it easier to later add middleware.